feat: add RBField
